Add support functions for A* heuristic function
Those two will help calculate the Manhattan Distance from each rock
to each switch. With this distance we can choose a minimum estimate of
distance from current state to goal state with minimum cost perfect
matching with a bipartite graph, in which rocks are one group and
switches the other.