Add heuristic function to fully implement A*
This is the first version of completed A* code.
This commit add a child class named AStar_state in order to store the
heuristic value (h) inside each state and alter the compare function to
use the value f = g + h for comparison.

The heuristic function itself work on the basis of minimum cost flow,
which solve a matching problem between two groups in a bipartite graph
so that the sum of cost of matching is minimum.

Currently there is one more issue regarding node count not correct
because the nodes in frontier (the 'open') priority queue is ignored.